Enterprise-HRMS/backend/requirements.txt
# Django & REST Framework
Django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.0

# Database
mysqlclient>=2.2

# Data Processing
pandas>=2.0

# Task Queue
celery>=5.3
redis>=4.5

# Utilities
python-dotenv>=1.0
